# api-specs/openapi-spec-20251021-190000.yaml
openapi: 3.1.0
info:
  title: Decentralized Autonomous Forum API
  version: 1.0.0
  description: |
    RESTful API for the Decentralized Autonomous Forum platform. Features:
    - Point economy with crypto rewards
    - User authentication (email + OAuth2)
    - Post/comment creation with IPFS storage
    - Reputation-based moderation
    - BNB Chain integration

    **Base URL**: `https://api.forum.example.com/v1`

    **Authentication**: Session-based (HttpOnly cookies) + JWT for API clients

    **Rate Limiting**:
    - Anonymous: 100 req/hour
    - Authenticated: 1000 req/hour
    - Authenticated (Level 3+): 5000 req/hour

  contact:
    name: API Support
    email: support@forum.example.com

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.forum.example.com/v1
    description: Production server
  - url: https://staging-api.forum.example.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie (HttpOnly, Secure, SameSite=Lax)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API clients

# ============================================================================
# REUSABLE SCHEMAS
# ============================================================================
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[A-Za-z0-9_-]+$'
          example: crypto_enthusiast
        display_name:
          type: string
          maxLength: 100
          example: Crypto Enthusiast
        bio:
          type: string
          maxLength: 500
          example: "Passionate about DeFi and BNB Chain"
        avatar_url:
          type: string
          format: uri
          example: "https://ipfs.io/ipfs/Qm..."
        points:
          type: integer
          minimum: 0
          example: 1250
        level:
          type: string
          enum: [new_user, active_user, trusted_user, moderator, senior_moderator, admin]
          example: active_user
        total_posts_created:
          type: integer
          example: 42
        total_comments_created:
          type: integer
          example: 128
        total_likes_received:
          type: integer
          example: 456
        is_verified:
          type: boolean
          example: false
        bnb_wallet_address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
        created_at:
          type: string
          format: date-time
          example: "2025-10-21T09:25:00Z"
      required:
        - id
        - username
        - points
        - level
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            recent_posts:
              type: array
              items:
                $ref: '#/components/schemas/PostSummary'
            recent_comments:
              type: array
              items:
                $ref: '#/components/schemas/CommentSummary'

    # ========== Post Schemas ==========
...
    PostSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        slug:
          type: string
        author:
          $ref: '#/components/schemas/UserSummary'
        like_count:
          type: integer
        comment_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    CommentSummary:
      type: object
      properties:
        id:
          type: integer
        post_id:
          type: integer
        body:
          type: string
        author:
          $ref: '#/components/schemas/UserSummary'
        like_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        level:
          type: string
        is_verified:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
      required:
        - error
        - message

    # ========== Pagination Schemas ==========
...
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Get authenticated user's profile with full details
      operationId: getCurrentUser
      security:
        - SessionAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}:
    get:
      tags:
        - Users
      summary: Get user profile by ID
      description: Get public profile of any user
      operationId: getUserById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}:
    get:
      tags:
        - Users
      summary: Get user profile by username
      description: Get public profile by username
      operationId: getUserByUsername
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found
...